Struct sudoxu::structs::cell::Cell

pub struct Cell(_);

A cell within the grid. Holds information about what candidates the cell has, and if the value of the cell is known.

Cell is a direct wrapper of a u16. In general, we think of Cells and u16s as being the same thing. That being said, not all operations on u16s are valid or safe in the context of a Cell’s representation. For this reason Cell does not implement Deref<u16>, nor is its inner u16 marked public. In many occasions it is still necessary to access the underyling u16. For this reason Cell contains an explicit to_u16 method for casting.

u16 Representation

0b  000000  0 0 0  0 0 0  0 0 0  0
    ^^^^^^  ^^^^^^^^^^^^^^^^^^^  ^
    |       |                    known bit
    |       candidates
    |      (9 8 7  6 5 4  3 2 1)
    banned bits

Known Bit

The first bit is the “known bit”. When it is set, the cell is assumed to have a known number (the cell has been filled out), as opposed to multiple candidates (having pencilmarks). This means there must only be a single “candidate bit” set.

// Valid u16 representation of a cell with value 9.
let cell_inner = 0b000000_100_000_000_1
// Invalid u16 representation. Known bit is set but candidate bits for 5 and 6 are set.
// This would represent a cell with a known value that is both 5 and 6.
let cell_inner = 0b000000_000_110_000_1
// Invalid u16 representation. Known bit is set but there are no set candidate bits.
// This would represent a cell with a known value, but no known value is set.
let cell_inner = 0b000000_000_000_000_1

Candidates

The following 9 bits are the “candidate” bits, these represent the possible candidate numbers 1..=9 that can go in the cell.

Banned Bits

The final 6 bits are unused in the current implementation. If they are not all set to zero, then the Cells representation is deemed invalid.

Safety

TODO

Nomenclature

We refer to a cell holding a known number as having a value of said number, while a cell not holding a known number as having candidates.

Cell(0b000000_100_000_000_1) has “value” 9.

Cell(0b000000_000_000_111_0) has “candidates” 1, 2, and 3.

Implementations

impl Cell

Cell Modification

pub unsafe fn set_known(&mut self)

Sets the known bit of the given cell.

Safety

Caller guarantees that the cell only has a single candidate, such that setting the known bit results in a valid representation.

pub unsafe fn set_candidates(&mut self, mask: CellMask)

Adds the set candidate bits in the CellMask to the Cell.

Safety

This function is provided to give a performance improvement to candidate removal. The caller guarantees that this operation respects the Cell representation.

pub fn remove_candidates(&mut self, mask: CellMask)

Removes the candidate bits set in the mask from the cell.

impl Cell

Base Implementations

pub fn check(self) -> Result<(), CandidateError>

Checks if a given Cell has an allowed representation.

pub fn new(n: u16) -> Result<Self, CandidateError>

Creates a new Cell from an inner u16. Returns a CandidateError if the u16’s does not follow the allowed representation.

pub unsafe fn new_unchecked(n: u16) -> Self

Creates a new Cell, which can either contain multiple candidates or be known. Does no checking to verify the representation of the u16 matches that of a Cell

Safety

The caller must ensure that the content of the u16 is a valid representation of a Cell.

pub fn new_known(n: Num) -> Cell

Creates a Cell with a known value from a Num.

pub fn to_u16(self) -> u16

Returns the inner u16 within the Cell.

pub fn is_known(self) -> bool

Returns whether the known bit is set.

pub unsafe fn zerod() -> Cell

Wrapper for Cell::new_unchecked(0)

Safety

This method intentionally returns an invalid Cell representation meant to be only used within internal calculations or accumulations. The caller guarantees that this form of a Cell is only used within such cases and is never returned or used as an actual Cell.

impl Cell

Finder

pub fn single_candidate(self) -> Option<Num>

Checks if a Cell is not known and only contains a single candidate.

Safety:

Does not check that the underyling u16 representation is correct. Any invalid Cell representation will return false.